# Performance optimizations for backups check